+++ R/dataset_methods.R
@@ -1368,6 +1368,73 @@ as_array_iterator <- function(dataset) {
 }
 
 
+#' Group windows of elements by key and reduce them
+#'
+#' @details This transformation maps each consecutive element in a dataset to a
+#' key using `key_func()` and groups the elements by key. It then applies
+#' `reduce_func()` to at most `window_size_func(key)` elements matching the same
+#' key. All except the final window for each key will contain
+#' `window_size_func(key)` elements; the final window may be smaller.
+#'
+#' You may provide either a constant `window_size` or a window size determined
+#' by the key through `window_size_func`.
+#'
+#' ````r
+#' window_size <-  5
+#' dataset <- range_dataset(to = 10) %>%
+#'   dataset_group_by_window(
+#'     key_func = function(x) x %% 2,
+#'     reduce_func = function(key, ds) dataset_batch(ds, window_size),
+#'     window_size = window_size
+#'   )
+#'
+#' it <- as_array_iterator(dataset)
+#' while (!is.null(elem <- iter_next(it)))
+#'   print(elem)
+#' #> tf.Tensor([0 2 4 6 8], shape=(5), dtype=int64)
+#' #> tf.Tensor([1 3 5 7 9], shape=(5), dtype=int64)
+#' ````
+#'
+#' @param dataset a TF Dataset
+#'
+#' @param key_func A function mapping a nested structure of tensors (having
+#'   shapes and types defined by `self$output_shapes` and `self$output_types`)
+#'   to a scalar `tf.int64` tensor.
+#'
+#' @param reduce_func A function mapping a key and a dataset of up to
+#'   `window_size` consecutive elements matching that key to another dataset.
+#'
+#' @param window_size A `tf.int64` scalar `tf.Tensor`, representing the number
+#'   of consecutive elements matching the same key to combine in a single batch,
+#'   which will be passed to `reduce_func`. Mutually exclusive with
+#'   `window_size_func`.
+#'
+#' @param window_size_func A function mapping a key to a `tf.int64` scalar
+#'   `tf.Tensor`, representing the number of consecutive elements matching the
+#'   same key to combine in a single batch, which will be passed to
+#'   `reduce_func`. Mutually exclusive with `window_size`.
+#'
+#' @param name (Optional.) A name for the Tensorflow operation.
+#'
+#' @seealso
+#'   +  <https://www.tensorflow.org/api_docs/python/tf/data/Dataset#group_by_window>
+#' @export
+dataset_group_by_window <-
+function(dataset, key_func, reduce_func,
+         window_size = NULL,
+         window_size_func = NULL,
+         name = NULL) {
+  if(!is.null(window_size_func))
+    window_size_func <- as_py_function(window_size_func)
+  as_tf_dataset(dataset$group_by_window(
+    key_func = as_py_function(key_func),
+    reduce_func = as_py_function(reduce_func),
+    window_size = as_integer_tensor(window_size),
+    window_size_func = window_size_func,
+    name = name
+  ))
+}
+
 
 #' Get the single element of the dataset.
 #'
@@ -1410,11 +1477,15 @@ as_array_iterator <- function(dataset) {
 #' expressed as TF Dataset operations, and you want to use those transformations
 #' while serving your model.
 #'
+#' @param x A TF Dataset
+#' @param name (Optional.) A name for the TensorFlow operation.
+#' @param ... passed on to `tensorflow::as_tensor()`
 #' @seealso
 #' +  <https://www.tensorflow.org/api_docs/python/tf/data/Dataset#get_single_element>
 #'
 #' @export
-#' @aliases get_single_element as_tensor
+#' @rdname as_tensor.tf_dataset
+#' @aliases get_single_element
 #' @importFrom tensorflow as_tensor
 as_tensor.tensorflow.python.data.ops.dataset_ops.DatasetV2 <- function(x, ..., name = NULL) {
   tensor <- x$get_single_element(name = name)
@@ -1424,10 +1495,10 @@ as_tensor.tensorflow.python.data.ops.dataset_ops.DatasetV2 <- function(x, ..., n
     tensor
 }
 
-#' @rdname as_tensor.tensorflow.python.data.ops.dataset_ops.DatasetV2
+#' @rdname as_tensor.tf_dataset
 #' @export
-as.array.tensorflow.python.data.ops.dataset_ops.DatasetV2 <- function(x)
-  as.array(as_tensor.tensorflow.python.data.ops.dataset_ops.DatasetV2(x))
+as.array.tensorflow.python.data.ops.dataset_ops.DatasetV2 <- function(x, ...)
+  as.array(as_tensor.tensorflow.python.data.ops.dataset_ops.DatasetV2(x, ...))
 
 
 #' @export