+++ NEWS.md
@@ -1,6 +1,8 @@
 # tfdatasets (development version)
 
--  New `dataset_unbatch()`
+- New `dataset_unbatch()`
+- New `as_tensor()` and `as.array()` methods which can be used on TF Datasets with a single element.
+
 
 # tfdatasets 2.7.0