+++ NEWS.md
@@ -1,6 +1,7 @@
 # tfdatasets (development version)
 
 - New `dataset_unbatch()`
+- New `dataset_group_by_window()`
 - New `as_tensor()` and `as.array()` methods which can be used on TF Datasets with a single element.