+++ NEWS.md
@@ -2,6 +2,7 @@
 
 - New `dataset_unbatch()`
 - New `dataset_group_by_window()`
+- New `dataset_take_while()`
 - New `as_tensor()` and `as.array()` methods which can be used on TF Datasets with a single element.